+++ python/acl_anthology/collections/eventindex.py
@@ -22,11 +22,14 @@ from typing import TYPE_CHECKING
 from ..containers import SlottedDict
 from ..text import MarkupText
 from ..utils.ids import AnthologyID, AnthologyIDTuple, parse_id
+from ..utils.logging import get_logger
 from .event import Event
 from .volume import Volume
 
 if TYPE_CHECKING:
     from ..anthology import Anthology
+
+log = get_logger()
 
 
 @define
@@ -47,7 +50,7 @@ class EventIndex(SlottedDict[Event]):
     reverse: dict[AnthologyIDTuple, set[str]] = field(
         init=False, repr=False, factory=lambda: defaultdict(set)
     )
-    is_data_loaded: bool = field(init=False, repr=False, default=False)
+    is_data_loaded: bool = field(init=False, repr=True, default=False)
 
     def by_volume(self, volume: Volume | AnthologyID) -> list[Event]:
         """Find events associated with a volume."""
@@ -61,6 +64,28 @@ class EventIndex(SlottedDict[Event]):
 
         return [self.data[event_id] for event_id in self.reverse[volume_fid]]
 
+    def reset(self) -> None:
+        """Resets the index."""
+        self.data = {}
+        self.reverse = defaultdict(set)
+        self.is_data_loaded = False
+
+    def _add_to_index(self, event: Event) -> None:
+        """Add an event to the index, adding already linked item IDs to it if applicable.
+
+        This function exists for internal use when creating new explicit events in a collection.  It should not be called manually.
+        """
+        if not self.is_data_loaded:
+            return
+
+        if event.id in self.data:
+            for co_id in self.data[event.id].colocated_ids:
+                if co_id not in event.colocated_ids:
+                    event.colocated_ids.append(co_id)
+        self.data[event.id] = event
+        for volume_fid in event.colocated_ids:
+            self.reverse[volume_fid].add(event.id)
+
     def load(self) -> None:
         """Load the entire Anthology data and build an index of events."""
         if self.is_data_loaded:
@@ -72,38 +97,46 @@ class EventIndex(SlottedDict[Event]):
             disable=(not self.verbose),
             description=" Building event index...",
         )
+        raised_exception = False
         for collection in iterator:
-            if (explicit_event := collection.get_event()) is not None:
-                if explicit_event.id in self.data:
-                    # This event has already been implicitly created in another file
-                    # See https://github.com/acl-org/acl-anthology/issues/2743#issuecomment-2453501562
-                    for co_id in self.data[explicit_event.id].colocated_ids:
-                        if co_id not in explicit_event.colocated_ids:
-                            explicit_event.colocated_ids.append(co_id)
-                self.data[explicit_event.id] = explicit_event
-                for volume_fid in explicit_event.colocated_ids:
-                    self.reverse[volume_fid].add(explicit_event.id)
+            try:
+                if (explicit_event := collection.get_event()) is not None:
+                    if explicit_event.id in self.data:
+                        # This event has already been implicitly created in another file
+                        # See https://github.com/acl-org/acl-anthology/issues/2743#issuecomment-2453501562
+                        for co_id in self.data[explicit_event.id].colocated_ids:
+                            if co_id not in explicit_event.colocated_ids:
+                                explicit_event.colocated_ids.append(co_id)
+                    self.data[explicit_event.id] = explicit_event
+                    for volume_fid in explicit_event.colocated_ids:
+                        self.reverse[volume_fid].add(explicit_event.id)
 
-            for volume in collection.volumes():
-                volume_fid = volume.full_id_tuple
-                if explicit_event is not None:
-                    self.reverse[volume_fid].add(explicit_event.id)
-                for venue_id in volume.venue_ids:
-                    event_id = f"{venue_id}-{volume.year}"
-                    if (event := self.data.get(event_id)) is None:
-                        # Implicitly create event if it doesn't exist yet
-                        venue_name = self.parent.venues[venue_id].name
-                        event_name = f"{venue_name} ({volume.year})"
-                        self.data[event_id] = Event(
-                            event_id,
-                            collection,
-                            is_explicit=False,
-                            colocated_ids=[volume_fid],
-                            title=MarkupText.from_string(event_name),
-                        )
-                    elif volume_fid not in event.colocated_ids:
-                        # Add implicit connection to existing event
-                        event.colocated_ids.append(volume_fid)
-                    self.reverse[volume_fid].add(event_id)
-
+                for volume in collection.volumes():
+                    volume_fid = volume.full_id_tuple
+                    if explicit_event is not None:
+                        self.reverse[volume_fid].add(explicit_event.id)
+                    for venue_id in volume.venue_ids:
+                        event_id = f"{venue_id}-{volume.year}"
+                        if (event := self.data.get(event_id)) is None:
+                            # Implicitly create event if it doesn't exist yet
+                            venue_name = self.parent.venues[venue_id].name
+                            event_name = f"{venue_name} ({volume.year})"
+                            self.data[event_id] = Event(
+                                event_id,
+                                collection,
+                                is_explicit=False,
+                                colocated_ids=[volume_fid],
+                                title=MarkupText.from_string(event_name),
+                            )
+                        elif volume_fid not in event.colocated_ids:
+                            # Add implicit connection to existing event
+                            event.colocated_ids.append(volume_fid)
+                        self.reverse[volume_fid].add(event_id)
+            except Exception as exc:
+                log.exception(exc)
+                raised_exception = True
+        if raised_exception:
+            raise Exception(
+                "An exception was raised while building EventIndex; check the logger for details."
+            )
         self.is_data_loaded = True